SYNERCHAT/app/backend/pyproject.toml
[project]
name = "synerchat-backend"
version = "0.1.0"
description = "FastAPI backend for Synerchat"
requires-python = ">=3.10"
dependencies = [
  "fastapi==0.112.2",
  "uvicorn[standard]==0.30.6",
  "pydantic==2.8.2",
  "python-jose[cryptography]==3.3.0",
  "websockets==12.0",
  "ecdsa==0.19.0",
  "eth-account==0.13.4",
  "eth-keys==0.5.0",
  "eth-utils==4.1.1",
  "web3==6.20.1",
  "python-multipart==0.0.9",
  "orjson==3.10.7",
  "slowapi==0.1.9"
]

[tool.uvicorn]
factory = false